Move socket helpers and context provider to ui-kit
