+++ packages/ui-kit/src/index.ts
@@ -1,1 +1,3 @@
+export * from './common/socket-context';
+export * from './lib/api';
 export * from './tools/create-frontend-tools';
