+++ packages/ui-kit/src/lib/api.ts
@@ -1,0 +1,1 @@
+export const API_BASE_URL = import.meta.env.VITE_OPS_BACKEND_URL || '';
